requirements.txt
telethon>=1.24.0
python-dotenv>=1.0.0